Histogram in python
Suppose, in a day the intervals given as  8:00 AM-12:00 PM,  12:00 PM-4.00 PM... total of 6 interval with 4 hour gap. In a array, the no of  mails a company received in these periods in every day of a week is given. Write a code showcasing the comparison of the mails received in each day in the given periods using histogram.